Hacked together some failure handling for MySQL
The way the program flow is right now, the Master agent checks the master
DB first, then the relay, then starts copying. Because of the way we have
MySQL set up as a master, there is a point in the program where if it
fails without completing, some data will never be copied over to the relay
because it resets the time range to copy. I added a column in the MySQL
CTIDtoTimestamp table that gets a 0 written to it when a line is inserted,
then in the cleanup init table (the last method in the Master.Run()
function) it writes a 1. I know this isn't great design, but until I have
time to rewrite more of it I don't want to mess with the program flow and
risk breaking our current working Tesla agents, so I'm leaving this one
until a different sprint.